refactor: Move episode lists to pages
