+++ src/pages/episodeslist/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import EpisodesList from './EpisodesList';
+
+function EpisodesListContainer(state) {
+  return <EpisodesList {...state} />;
+}
+
+export default EpisodesListContainer;
