+++ src/reducers/episodes.js
@@ -2,6 +2,8 @@ import * as actions from 'app/actions/episodes';
 
 const initialState = {
   episodes: [],
+  page: 1,
+  per_page: 30,
   total: 0,
 }
